mediabox: Center the currently selected item in the navbar
The mediabox displays the thumbnails of the items in the gallery at the
bottom of the page when viewing the mediabox. If you had a lot of items,
you'd only see the ones that fit in your viewing width.

This change centers the currently selected item in the middle of the
navbar.